requirements.txt
asgiref==3.8.1
boto3==1.37.32
botocore==1.37.32
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
Django==5.1.7
django-allauth==0.54.0
django-countries==7.6.1
django-crispy-forms==2.3
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python3-openid==3.2.0
requests-oauthlib==2.0.0
s3transfer==0.11.4
sqlparse==0.5.3
whitenoise==6.9.0
